Renaming
- Renamed constants.rs to config.rs
- Renamed config.rs to utilities.rs
