--- src/application/constants.rs
@@ -1,2 +1,0 @@
-pub static VERSION: &str = "0.1.1";
-pub static APPLICATION_ID: &str = "dev.edfloreshz.Done";
