+++ src/application/mod.rs
@@ -1,6 +1,6 @@
 pub mod application;
+pub mod utilities;
 pub mod config;
-pub mod constants;
 pub mod localize;
 pub mod resources;
 pub mod window;
